refactor: rename type file to "app" and remove imports from namespaced
